requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
Werkzeug==2.3.7

web: gunicorn app:app
gunicorn==21.2.0